requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
dataclasses-json==0.6.7
Deprecated==1.2.14
dirtyjson==1.0.8
distlib==0.3.8
distro==1.9.0
dj-rest-auth==6.0.0
Django==5.1.1
django-allauth==64.2.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.16.0
filetype==1.2.0
frozenlist==1.4.1
fsspec==2024.9.0
ftfy==6.2.3
greenlet==3.1.1
grpcio==1.66.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.1
identify==2.6.1
idna==3.10
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.6
langchain-openai==0.2.0
langsmith==0.1.128
llama-index-core==0.11.13.post1
llama-index-readers-pdf-marker==0.2.0
marker-pdf==0.2.17
MarkupSafe==2.1.5
marshmallow==3.22.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
nodeenv==1.9.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.48.0
opencv-python==4.10.0.84
orjson==3.10.7
packaging==24.1
pdftext==0.3.10
pillow==10.4.0
platformdirs==4.3.3
pre-commit==3.8.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
PyJWT==2.9.0
pypdfium2==4.30.0
python-dotenv==1.0.1
PyYAML==6.0.2
RapidFuzz==3.10.0
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
safetensors==0.4.5
scikit-learn==1.4.2
scipy==1.14.1
setuptools==75.1.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlparse==0.5.1
surya-ocr==0.5.0
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
texify==0.1.10
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.20.0
torch==2.4.1
tqdm==4.66.5
transformers==4.45.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.26.4
wcwidth==0.2.13
wrapt==1.16.0
yarl==1.12.1
